chore: extract jc2m into multiple files
